+++ src/com/rb2750/lwjgl/util/Vector2.java
@@ -8,8 +8,8 @@ import lombok.Setter;
 public class Vector2 {
     @Getter
     @Setter
-    private float x;
+    private double x;
     @Getter
     @Setter
-    private float y;
+    private double y;
 }
